Make TM params more resistent to variance in param
Fixes #10

Steps' 16 value was right at the extreme, meaning it would only be hit
if a pot could actually be read at, or above, 1.
Prob used a threshold of 0.001 for locking, which was too small. Now, it
narrows the range of the param by 0.02 such that you can now hit
theoretically any value, but its locking is more robust .